requirements.txt
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
binaryornot==0.4.4
bitcoinlib==0.6.15
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cookiecutter==2.6.0
crashtest==0.4.1
distlib==0.3.8
dulwich==0.21.7
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
fastecdsa==2.3.2
fastjsonschema==2.19.1
filelock==3.13.3
flet==0.21.2
flet-core==0.21.2
flet-runtime==0.21.2
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
installer==0.7.0
jaraco.classes==3.4.0
Jinja2==3.1.3
keyring==24.3.1
libusb1==3.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.8
ndeflib==0.3.3
nfcpy==1.0.4
numpy==1.26.4
oauthlib==3.2.2
opencv-python-headless==4.9.0.80
packaging==23.2
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.10.0
platformdirs==4.2.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
ptyprocess==0.7.0
pyaes==1.6.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pyDes==2.0.1
Pygments==2.17.2
pyperclip==1.8.2
pypng==0.20220715.0
pyproject_hooks==1.0.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
pyzbar==0.1.9
qrcode==7.4.2
rapidfuzz==3.8.1
repath==0.9.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
scrypt==0.8.24
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.12.4
trove-classifiers==2024.3.25
types-python-dateutil==2.8.19.20240311
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
virtualenv==20.25.1
watchdog==4.0.0
watchfiles==0.21.0
websockets==12.0
xattr==1.1.0
zipp==3.18.1
